test,fix(storage): add integrations tests for FileController

+++ src/main/java/ru/masnaviev/cloudfile/service/impl/S3FileServiceImpl.java
@@ -1,9 +1,7 @@
 package ru.masnaviev.cloudfile.service.impl;
 
 import io.minio.GetObjectResponse;
-import io.minio.Result;
 import io.minio.StatObjectResponse;
-import io.minio.errors.MinioException;
 import io.minio.messages.DeleteObject;
 import io.minio.messages.Item;
 import lombok.RequiredArgsConstructor;
@@ -16,17 +14,18 @@ import ru.masnaviev.cloudfile.exception.resource.*;
 import ru.masnaviev.cloudfile.repository.MinioRepository;
 import ru.masnaviev.cloudfile.service.S3FileService;
 import ru.masnaviev.cloudfile.util.Resource;
-import ru.masnaviev.cloudfile.util.ResourceBuilder;
+import ru.masnaviev.cloudfile.util.ResourceType;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import static ru.masnaviev.cloudfile.constatnts.ErrorMessages.*;
+import static ru.masnaviev.cloudfile.util.ResourceBuilder.createFrom;
 import static ru.masnaviev.cloudfile.util.ResourceType.DIRECTORY;
 import static ru.masnaviev.cloudfile.util.ResourceType.FILE;
 import static ru.masnaviev.cloudfile.util.ZipBuilder.createZipFromResources;
@@ -38,19 +37,27 @@ public class S3FileServiceImpl implements S3FileService {
     private final MinioRepository repository;
 
     public ResourceInfoResponse getResourceInfo(Long userId, String path) {
-        Resource resourceData = ResourceBuilder.createFrom(userId, path);
+        Resource resourceData = createFrom(userId, path);
 
         checkPathExists(resourceData.getPathWithoutResourceName());
-        checkResourceExists(resourceData);
+        checkResourceExists(resourceData.getFullPath(), resourceData.getResourceType());
 
-        return getResourceInfo(resourceData);
+
+        StatObjectResponse response = repository.getResourceInfo(resourceData.getFullPath());
+
+        return buildResponse(resourceData.getPathWithoutUsernameAndResourceName(), resourceData.getResourceName(),
+                resourceData.getResourceType() == FILE ? response.size() : null, resourceData.getResourceType());
     }
 
     public void deleteResource(Long userId, String path) {
-        Resource resourceData = ResourceBuilder.createFrom(userId, path);
+        Resource resourceData = createFrom(userId, path);
+
+        if (path.equals("/")) {
+            throw new ParentDirectoryDeletionException(PROTECTED_PARENT_DIRECTORY);
+        }
 
         checkPathExists(resourceData.getPathWithoutResourceName());
-        checkResourceExists(resourceData);
+        checkResourceExists(resourceData.getFullPath(), resourceData.getResourceType());
 
         if (resourceData.getResourceType() == DIRECTORY)
             deleteDirectory(resourceData);
@@ -59,10 +66,10 @@ public class S3FileServiceImpl implements S3FileService {
     }
 
     public DownloadResourceResponse downloadResource(Long userId, String path) {
-        Resource resourceData = ResourceBuilder.createFrom(userId, path);
+        Resource resourceData = createFrom(userId, path);
 
         checkPathExists(resourceData.getPathWithoutResourceName());
-        checkResourceExists(resourceData);
+        checkResourceExists(resourceData.getFullPath(), resourceData.getResourceType());
 
         InputStreamResource resource = resourceData.getResourceType() == DIRECTORY ?
                 downloadDirectory(resourceData) :
@@ -71,9 +78,126 @@ public class S3FileServiceImpl implements S3FileService {
         return new DownloadResourceResponse(resourceData.getResourceName(), resource, resourceData.getResourceType());
     }
 
+    public ResourceInfoResponse uploadDirectory(Long userId, String path) {
+        Resource resourceData = createFrom(userId, path);
+
+        checkResourceExists(resourceData.getPathWithoutResourceName(), resourceData.getResourceType());
+        checkPathEndWithSlash(resourceData);
+
+        if (repository.checkResourceExists(resourceData.getFullPath())) {
+            throw new DirectoryAlreadyExistsException(DIRECTORY_ALREADY_EXISTS);
+        }
+
+        repository.uploadDirectory(resourceData.getFullPath());
+
+        return buildResponse(resourceData.getPathWithoutUsernameAndResourceName(), resourceData.getResourceName(), null, DIRECTORY);
+    }
+
+    public List<ResourceInfoResponse> getDirectoryContentsInfo(Long userId, String path) {
+        Resource resourceData = createFrom(userId, path);
+
+        checkPathExists(resourceData.getPathWithoutResourceName());
+
+        Map<String, Item> resultMap = repository.getResourcesItemsByPrefix(resourceData.getFullPath(), false);
+
+        List<ResourceInfoResponse> responses = new ArrayList<>();
+
+        for (Map.Entry<String, Item> result : resultMap.entrySet()) {
+
+            var resultData = createFrom(userId,
+                    result.getKey().replace(resourceData.getUserFolder(), ""));
+
+            ResourceInfoResponse response = buildResponse(resultData.getPathWithoutUsernameAndResourceName(), resultData.getResourceName(),
+                    resultData.getResourceType() == DIRECTORY ? null : result.getValue().size(), resultData.getResourceType());
+
+            responses.add(response);
+
+        }
+        return responses;
+    }
+
+    public List<ResourceInfoResponse> uploadResources(Long userId, String path, List<MultipartFile> files) {
+        Resource pathData = createFrom(userId, path);
+        checkPathExists(pathData.getFullPath());
+
+        checkPathEndWithSlash(pathData);
+
+        List<ResourceInfoResponse> responses = new ArrayList<>();
+
+        for (var file : files) {
+            Resource resourceData = createFrom(userId, path + file.getOriginalFilename());
+
+            if (repository.checkResourceExists(resourceData.getFullPath())) {
+                throw new FileAlreadyExistsException(FILE_ALREADY_EXIST);
+            }
+
+            List<String> pathsToCreate = resourceData.getPathsList();
+
+            Set<String> existedPaths = repository.getResourcesItemsByPrefix(pathData.getFullPath(), true).keySet();
+
+            pathsToCreate.removeAll(existedPaths);
+
+            for (String pathToCreate : pathsToCreate) {
+                repository.uploadDirectory(pathToCreate);
+                var normalizedData = createFrom(userId, pathToCreate);
+                ResourceInfoResponse response = buildResponse(normalizedData.getPathWithoutUsernameAndResourceName(), normalizedData.getResourceName(), null, normalizedData.getResourceType());
+
+                responses.add(response);
+            }
+
+            repository.uploadFile(resourceData.getFullPath(), file);
+            ResourceInfoResponse response = buildResponse(resourceData.getPathWithoutUsernameAndResourceName(), resourceData.getResourceName(), file.getSize(), resourceData.getResourceType());
+
+            responses.add(response);
+        }
+        return responses;
+    }
+
+    public void createUserDirectory(Long userId) {
+        String userFolder = "user-" + userId + "-files" + "/";
+        repository.uploadDirectory(userFolder);
+    }
+
+    private void deleteDirectory(Resource resourceData) {
+
+        Map<String, Item> ResourcesItems = repository.getResourcesItemsByPrefix(resourceData.getFullPath(), true);
+
+        List<DeleteObject> objectsForDelete = ResourcesItems.keySet().stream().map(DeleteObject::new).toList();
+
+        repository.deleteResources(objectsForDelete);
+    }
+
+    private void checkPathExists(String path) {
+        if (!repository.checkResourceExists(path)) {
+            throw new PathNotFoundException(PATH_NOT_FOUND);
+        }
+    }
+
+    private void checkResourceExists(String path, ResourceType resourceType) {
+        if (!repository.checkResourceExists(path)) {
+            throw new ResourceNotFoundException(resourceType == DIRECTORY ?
+                    DIRECTORY_NOT_FOUND :
+                    FILE_NOT_FOUND);
+        }
+    }
+
+    private void checkPathEndWithSlash(Resource pathData) {
+        if (!pathData.getFullPath().endsWith("/")) {
+            throw new PathMustEndWithSlashException(PATH_MUST_BE_END_SLASH);
+        }
+    }
+
+    private ResourceInfoResponse buildResponse(String path, String name, Long size, ResourceType resourceType) {
+        return ResourceInfoResponse.builder()
+                .path(path)
+                .name(name)
+                .size(size)
+                .resourceType(resourceType)
+                .build();
+    }
+
     private InputStreamResource downloadDirectory(Resource resourceData) {
-        Set<String> resourcesForDownload = toItemMapByPath(repository
-                .getResourcesByPrefix(resourceData.getFullPath(), true)).keySet();
+        Set<String> resourcesForDownload = repository.getResourcesItemsByPrefix(resourceData.getFullPath(), true).keySet();
 
         Map<String, GetObjectResponse> downloadedResources = resourcesForDownload
                 .stream()
@@ -86,169 +210,5 @@ public class S3FileServiceImpl implements S3FileService {
 
     private InputStreamResource downloadFile(Resource resourceData) {
         return new InputStreamResource(repository.downloadResource(resourceData.getFullPath()));
-    }
-
-    public ResourceInfoResponse uploadDirectory(Long userId, String path) {
-        Resource resourceData = ResourceBuilder.createFrom(userId, path);
-
-        checkPathExists(resourceData.getPathWithoutResourceName());
-
-        if (!resourceData.getFullPath().endsWith("/")) {
-            throw new PathMustEndWithSlashException(PATH_MUST_BE_END_SLASH);
-        }
-        if (repository.checkResourceExists(resourceData.getFullPath())) {
-            throw new DirectoryAlreadyExistsException(DIRECTORY_ALREADY_EXISTS);
-        }
-
-        repository.uploadDirectory(resourceData.getFullPath());
-
-        return ResourceInfoResponse.builder()
-                .path(resourceData.getPathWithoutUsernameAndResourceName())
-                .name(resourceData.getResourceName())
-                .size(null)
-                .resourceType(DIRECTORY)
-                .build();
-    }
-
-    public List<ResourceInfoResponse> getDirectoryContentsInfo(Long userId, String path) {
-        Resource resourceData = ResourceBuilder.createFrom(userId, path);
-
-        checkPathExists(resourceData.getPathWithoutResourceName());
-
-        Iterable<Result<Item>> results = repository.getResourcesByPrefix(resourceData.getFullPath(), false);
-
-        List<ResourceInfoResponse> responses = new ArrayList<>();
-
-        Map<String, Item> resultMap = toItemMapByPath(results);
-
-        for (Map.Entry<String, Item> result : resultMap.entrySet()) {
-
-            if (result.getKey().equals(resourceData.getFullPath()))
-                responses.add(toResponse(userId, result.getKey(), result.getValue().size(), resourceData));
-
-        }
-        return responses;
-    }
-
-    public List<ResourceInfoResponse> uploadResources(Long userId, String path, List<MultipartFile> files) {
-
-        // Проверяем, что путь до папки, куда будут загружаться ресурсы, существует
-        var pathData = ResourceBuilder.createFrom(userId, path);
-        if (!repository.checkResourceExists(pathData.getFullPath())) {
-            throw new PathNotFoundException(PATH_NOT_FOUND);
-        }
-
-        // Проверка, что путь кончается на "/"
-        if (!pathData.getFullPath().endsWith("/")) {
-            throw new PathMustEndWithSlashException(PATH_MUST_BE_END_SLASH);
-        }
-
-        List<ResourceInfoResponse> responses = new ArrayList<>();
-
-        for (var file : files) {
-            Resource resourceData = ResourceBuilder.createFrom(userId, path + "/" + file.getOriginalFilename());
-
-            if (repository.checkResourceExists(resourceData.getFullPath())) {
-                throw new FileAlreadyExistsException(FILE_ALREADY_EXIST);
-            }
-
-            // Получаем список папок для файла
-            List<String> pathsToCreate = resourceData.getPathsList();
-
-            // Рекурсивно получаем содержимое папки
-            Iterable<Result<Item>> existedResources = repository.getResourcesByPrefix(pathData.getFullPath(), true);
-
-            // Получаем список существующих ресурсов
-            Set<String> existedPaths = toItemMapByPath(existedResources).keySet();
-
-            pathsToCreate.removeAll(existedPaths);
-
-            // Создаем папки при необходимости
-            for (String pathToCreate : pathsToCreate) {
-                repository.uploadDirectory(pathToCreate);
-                var normalizedData = ResourceBuilder.createFrom(userId, pathToCreate);
-                responses.add(toResponse(userId, pathToCreate, null, normalizedData));
-            }
-
-            repository.uploadFile(resourceData.getFullPath(), file);
-
-            ResourceInfoResponse fileUploadResponse = toResponse(userId, resourceData.getFullPath(), file.getSize(), resourceData);
-
-            responses.add(fileUploadResponse);
-        }
-        return responses;
-    }
-
-
-    private ResourceInfoResponse getResourceInfo(Resource resourceData) {
-
-        StatObjectResponse response = repository.getResourceInfo(resourceData.getFullPath());
-
-        return ResourceInfoResponse.builder()
-                .path(resourceData.getPathWithoutUsernameAndResourceName())
-                .name(resourceData.getResourceName())
-                .size(resourceData.getResourceType() == FILE ? response.size() : null)
-                .resourceType(resourceData.getResourceType())
-                .build();
-    }
-
-    private void deleteDirectory(Resource resourceData) {
-
-        Iterable<Result<Item>> resourcesForDelete = repository.getResourcesByPrefix(resourceData.getFullPath(), true);
-
-        Set<String> pathsForDelete = toItemMapByPath(resourcesForDelete).keySet();
-
-        List<DeleteObject> objectsForDelete = pathsForDelete.stream().map(DeleteObject::new).toList();
-
-        repository.deleteResources(objectsForDelete);
-    }
-
-    private ResourceInfoResponse toResponse(Long userId, String objectName, Long size,
-                                            Resource resourceData) {
-
-        var resultData = ResourceBuilder.createFrom(userId,
-                objectName.replace(resourceData.getUserFolder(), ""));
-
-        return ResourceInfoResponse.builder()
-                .path(resultData.getPathWithoutUsernameAndResourceName())
-                .name(resultData.getResourceName())
-                .size(resultData.getResourceType() == DIRECTORY ? null : size)
-                .resourceType(resultData.getResourceType())
-                .build();
-    }
-
-    private Map<String, Item> toItemMapByPath(Iterable<Result<Item>> results) {
-        Map<String, Item> paths = new HashMap<>();
-
-        try {
-            for (Result<Item> result : results) {
-                paths.put(result.get().objectName(), result.get());
-            }
-        } catch (MinioException | NoSuchAlgorithmException | InvalidKeyException | IOException e) {
-            throw new MinioOperationException(e.getMessage(), e.getCause());
-        }
-        return paths;
-    }
-
-    public void createUserDirectory(Long userId) {
-
-        //TODO дублируется folder в NormalizedResourceData, подумать как отрефакторить
-        //TODO сделать так чтобы корневую папку нельзя было удалить
-        String userFolder = "user-" + userId + "-files" + "/";
-        repository.uploadDirectory(userFolder);
-    }
-
-    private void checkPathExists(String path) {
-        if (!repository.checkResourceExists(path)) {
-            throw new PathNotFoundException(ru.masnaviev.cloudfile.constatnts.ErrorMessages.PATH_NOT_FOUND);
-        }
-    }
-
-    private void checkResourceExists(Resource resourceData) {
-        if (!repository.checkResourceExists(resourceData.getFullPath())) {
-            throw new ResourceNotFoundException(resourceData.getResourceType() == DIRECTORY ?
-                    DIRECTORY_NOT_FOUND :
-                    FILE_NOT_FOUND);
-        }
     }
 }
